Move magento-customer-account & magento-customer-order into magento-customer package
