+++ .changeset/giant-yaks-brake.md
@@ -1,0 +1,8 @@
+---
+'@graphcommerce/magento-graphcms': patch
+'@graphcommerce/magento-customer': patch
+'@graphcommerce/magento-customer-account': patch
+'@graphcommerce/magento-customer-order': patch
+---
+
+Move magento-customer-account & magento-customer-order into magento-customer package (magento-customer-account & magento-customer-order are now deprecated)
