--- packages/magento-customer-account/components/index.ts
@@ -1,5 +1,0 @@
-export * from './AccountLatestOrder/AccountLatestOrder'
-export * from './AccountMenu/AccountMenu'
-export * from './AccountMenuItem/AccountMenuItem'
-export * from './AccountSignInUpForm/AccountSignInUpForm'
-export * from './AccountOrders/AccountOrders'
