+++ packages/magento-customer-account/index.ts
@@ -1,4 +1,8 @@
-export * from './components'
-export * from './queries/AccountDashboardOrders.gql'
-export * from './queries/AccountDashboardAddresses.gql'
-export * from './queries/AccountDashboard.gql'
+export * from '@graphcommerce/magento-customer/components/AccountLatestOrder/AccountLatestOrder'
+export * from '@graphcommerce/magento-customer/components/AccountMenu/AccountMenu'
+export * from '@graphcommerce/magento-customer/components/AccountMenuItem/AccountMenuItem'
+export * from '@graphcommerce/magento-customer/components/AccountSignInUpForm/AccountSignInUpForm'
+export * from '@graphcommerce/magento-customer/components/AccountOrders/AccountOrders'
+export * from '@graphcommerce/magento-customer/graphql/AccountDashboardOrders.gql'
+export * from '@graphcommerce/magento-customer/graphql/AccountDashboardAddresses.gql'
+export * from '@graphcommerce/magento-customer/graphql/AccountDashboard.gql'
