--- packages/magento-customer-order/components/index.ts
@@ -1,9 +1,0 @@
-export * from './NoOrdersFound/NoOrdersFound'
-export * from './OrderCard/OrderCard'
-export * from './OrderCardItem/OrderCardItem'
-export * from './OrderCardItemImage/OrderCardItemImage'
-export * from './OrderDetails/OrderDetails'
-export * from './OrderItem/OrderItem'
-export * from './OrderItems/OrderItems'
-export * from './OrderStateLabel/OrderStateLabel'
-export * from './OrderStateLabel/OrderStateLabelInline'
