--- packages/magento-customer-order/hooks/index.ts
@@ -1,4 +1,0 @@
-export { default as useOrderCardItemImages } from './useOrderCardItemImages'
-export * from './UseOrderCardItemImages.gql'
-export * from './OrderCardItemImages.gql'
-export * from './OrderCardItemImage.gql'
