+++ packages/magento-customer/components/AccountLatestOrder/AccountLatestOrder.tsx
@@ -1,12 +1,9 @@
-import {
-  NoOrdersFound,
-  OrderCard,
-  useOrderCardItemImages,
-} from '@graphcommerce/magento-customer-order'
-
 import { SectionContainer } from '@graphcommerce/next-ui'
 import React from 'react'
+import useOrderCardItemImages from '../../hooks/useOrderCardItemImages'
 import { AccountOrdersFragment } from '../AccountOrders/AccountOrders.gql'
+import { NoOrdersFound } from '../NoOrdersFound/NoOrdersFound'
+import { OrderCard } from '../OrderCard/OrderCard'
 
 export type AccountLatestOrderProps = AccountOrdersFragment & {
   loading: boolean
