+++ packages/magento-customer/components/OrderCardItemImage/OrderCardItemImage.tsx
