+++ packages/magento-customer/components/index.ts
@@ -1,6 +1,11 @@
 export * from './AccountAddress/AccountAddress'
 export * from './AccountAddresses/AccountAddresses.gql'
 export * from './AccountAddresses/AccountAddresses'
+export * from './AccountLatestOrder/AccountLatestOrder'
+export * from './AccountMenu/AccountMenu'
+export * from './AccountMenuItem/AccountMenuItem'
+export * from './AccountOrders/AccountOrders'
+export * from './AccountSignInUpForm/AccountSignInUpForm'
 export * from './AddressFields/AddressFields'
 export * from './AddressMultiLine/AddressMultiLine'
 export * from './AddressSingleLine/AddressSingleLine'
@@ -16,6 +21,15 @@ export * from './EditAddressForm/EditAddressForm'
 export * from './ForgotPasswordForm/ForgotPassword.gql'
 export * from './ForgotPasswordForm/ForgotPasswordForm'
 export * from './NameFields/NameFields'
+export * from './NoOrdersFound/NoOrdersFound'
+export * from './OrderCard/OrderCard'
+export * from './OrderCardItem/OrderCardItem'
+export * from './OrderCardItemImage/OrderCardItemImage'
+export * from './OrderDetails/OrderDetails'
+export * from './OrderItem/OrderItem'
+export * from './OrderItems/OrderItems'
+export * from './OrderStateLabel/OrderStateLabel'
+export * from './OrderStateLabel/OrderStateLabelInline'
 export * from './ResetPasswordForm/ResetPasswordForm'
 export * from './SignInForm/SignIn.gql'
 export * from './SignInForm/SignInForm'
