+++ packages/magento-customer/hooks/OrderCardItemImage.graphql
