+++ packages/magento-customer/hooks/index.ts
@@ -6,3 +6,7 @@ export * from './useCustomerQuery'
 export * from './useCustomerSession'
 export * from './useFormIsEmailAvailable'
 export * from './useGuestQuery'
+export { default as useOrderCardItemImages } from './useOrderCardItemImages'
+export * from './UseOrderCardItemImages.gql'
+export * from './OrderCardItemImages.gql'
+export * from './OrderCardItemImage.gql'
