+++ packages/magento-customer/index.ts
@@ -2,3 +2,7 @@ export * from './components'
 export * from './hooks'
 export * from './typePolicies'
 export * from './link/createCustomerTokenLink'
+export * from './graphql/AccountDashboardOrders.gql'
+export * from './graphql/AccountDashboardAddresses.gql'
+export * from './graphql/AccountDashboard.gql'
+export * from './graphql/OrderDetailPage.gql'
